test: added orderitems and review resource tests

+++ shop/src/main/java/me/shail/services/ReviewService.java
@@ -20,6 +20,8 @@ import java.util.UUID;
 @ApplicationScoped
 public class ReviewService {
 
+    public final static String REVIEW_NOT_EXIST_ERROR_MSG = "Review does not exist. Id: ";
+
     @Inject
     ReviewRepository reviewRepository;
 
@@ -38,25 +40,36 @@ public class ReviewService {
     @WithCustomStatelessSession
     public Uni<Long> countReviewsByProductId(UUID productId) {
         log.debug("Request to get count of Reviews for Product Id: {}", productId);
-        return this.reviewRepository.countReviewsByProductId(productId);
+        return ProductService.generateUni_ExistById(this.productRepository, productId, false)
+                .chain(exists -> {
+                    if (!exists)
+                        return Uni.createFrom().failure(() ->
+                                new EntityNotFoundException(ProductService.getProductNotExistErrorMsg(productId))
+                        );
+                    return this.reviewRepository.countReviewsByProductId(productId);
+                });
     }
 
     @WithCustomStatelessSession
     public Uni<List<ReviewDto>> findReviewsByProductId(UUID productId) {
         log.debug("Request to get all Reviews");
-        return this.reviewRepository.findReviewsByProductId(productId)
-                .map(reviews -> reviews.stream()
-                        .map(ReviewService::mapToDto)
-                        .toList());
+        return ProductService.generateUni_ExistById(this.productRepository, productId, false)
+                .chain(exists -> {
+                    if (!exists)
+                        return Uni.createFrom().failure(() ->
+                                new EntityNotFoundException(ProductService.getProductNotExistErrorMsg(productId))
+                        );
+                    return this.reviewRepository.findReviewsByProductId(productId)
+                            .map(reviews -> reviews.stream()
+                                    .map(ReviewService::mapToDto)
+                                    .toList());
+                });
     }
 
     @WithCustomStatelessSession
     public Uni<ReviewDto> findById(UUID reviewId) {
         log.debug("Request to get Review: {}", reviewId);
-        return this.reviewRepository.findById(reviewId)
-                .onItem().ifNull().failWith(() ->
-                        new EntityNotFoundException("Review does not exist. Id: " + reviewId)
-                )
+        return generateUni_FindById(this.reviewRepository, reviewId, false)
                 .map(ReviewService::mapToDto);
     }
 
@@ -69,7 +82,7 @@ public class ReviewService {
                 .chain(exists -> {
                     if (!exists)
                         return Uni.createFrom().failure(() ->
-                                new EntityNotFoundException("Product does not exist. ID: " + productId)
+                                new EntityNotFoundException(ProductService.getProductNotExistErrorMsg(productId))
                         );
 
                     return this.reviewRepository
@@ -87,12 +100,47 @@ public class ReviewService {
     @WithTransaction
     public Uni<Boolean> deleteById(UUID reviewId) {
         log.debug("Request to delete Review: {}", reviewId);
-        return this.reviewRepository.deleteById(reviewId);
+        return generateUni_ExistById(this.reviewRepository, reviewId, true)
+                .chain(exists -> {
+                    if (!exists)
+                        return Uni.createFrom().failure(()
+                                -> new EntityNotFoundException(getReviewNotExistErrorMsg(reviewId))
+                        );
+
+                    return this.reviewRepository.deleteById(reviewId);
+                });
     }
 
     @WithTransaction
     public Uni<Integer> deleteAllReviewsByProductId(UUID productId) {
         log.debug("Request to delete all reviews for Product Id: {}", productId);
-        return this.reviewRepository.deleteReviewsByProductId(productId);
+        return ProductService.generateUni_ExistById(this.productRepository, productId, true)
+                .chain(exists -> {
+                    if (!exists)
+                        return Uni.createFrom().failure(() ->
+                                new EntityNotFoundException(ProductService.getProductNotExistErrorMsg(productId))
+                        );
+                    return this.reviewRepository.deleteReviewsByProductId(productId);
+                });
+    }
+
+    public static Uni<Review> generateUni_FindById(ReviewRepository repository, UUID reviewId, boolean managed) {
+        Uni<Review> generatedUni = managed
+                ? repository.findByIdManaged(reviewId)
+                : repository.findByIdStateless(reviewId);
+
+        return generatedUni.onItem().ifNull().failWith(() ->
+                new EntityNotFoundException(getReviewNotExistErrorMsg(reviewId))
+        );
+    }
+
+    public static Uni<Boolean> generateUni_ExistById(ReviewRepository repository, UUID reviewId, boolean managed) {
+        return managed
+                ? repository.existByIdManaged(reviewId)
+                : repository.existByIdStateless(reviewId);
+    }
+
+    public static String getReviewNotExistErrorMsg(UUID reviewId) {
+        return REVIEW_NOT_EXIST_ERROR_MSG + reviewId;
     }
 }
